Fixes for hafla: repi part and hook break
It was checked at TNM 2025 with a member of an authoring band
that this is the correct version.

Change to the repi part is that hits and rims are swapped, to match
how the tune is written in the player.

Change to the hook break is that first "A" response is supposed to
happen 1 16th later than was written.